+++ resources/js/modals/MediaViewModal.vue
@@ -14,9 +14,9 @@
           <div class="o1-flex">
             <!-- File info and media fields -->
             <div
-              class="o1-flex o1-flex-col o1-pr-4 o1-border-r o1-border-slate-200 o1-mr-4 o1-max-w-sm o1-w-full dark:o1-border-slate-700"
+              class="o1-flex o1-flex-col o1-pr-4 o1-border-r o1-border-gray-200 o1-mr-4 o1-max-w-sm o1-w-full dark:o1-border-gray-700"
             >
-              <div class="border-b o1-border-slate-200 dark:o1-border-slate-700 o1-mb-5">
+              <div class="border-b o1-border-gray-200 dark:o1-border-gray-700 o1-mb-5">
                 <MediaViewModalInfoListItem :label="__('novaMediaHub.viewModalIdTitle')" :value="mediaItem.id" />
                 <MediaViewModalInfoListItem :label="__('novaMediaHub.fileNameTitle')" :value="mediaItem.file_name" />
                 <MediaViewModalInfoListItem :label="__('novaMediaHub.fileSizeTitle')" :value="fileSize" />
@@ -44,7 +44,7 @@
             >
               <img
                 v-if="type === 'image'"
-                class="o1-object-contain o1-max-w-full o1-w-full o1-max-h-full o1-border o1-border-slate-100 o1-bg-slate-50 o1-min-h-0 dark:o1-bg-slate-900 dark:o1-border-slate-700"
+                class="o1-object-contain o1-max-w-full o1-w-full o1-max-h-full o1-border o1-border-gray-100 o1-bg-gray-50 o1-min-h-0 dark:o1-bg-gray-900 dark:o1-border-gray-700"
                 :src="mediaItem.url"
                 :alt="mediaItem.file_name"
               />